% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffp_evd_simu_loc_demo.R
\name{ffp_evd_simu_loc_demo_main}
\alias{ffp_evd_simu_loc_demo_main}
\title{simulate demographics distributions by location}
\usage{
ffp_evd_simu_loc_demo_main(
  it_M_location = 4,
  ar_it_N_pop_groups = c(3, 2),
  fl_sd_bound = 1,
  ar_st_outer_groups = c("lower_edu", "middle_edu", "higher_edu"),
  ar_st_dims = c("education", "age_group"),
  ar_outer_group_share = c(0.5, 0.25, 0.25),
  ar_fl_withingrp_runif_min = c(0.3, 0.2, 0.1),
  ar_fl_withingrp_runif_max = c(0.4, 0.8, 0.2),
  bl_inner_is_age = TRUE,
  it_age_group_year_span = 25,
  it_age_youngest_1st_group = 0
)
}
\arguments{
\item{it_m_location}{integer the number of locations}
}
\value{
an array of tax-liabilities for particular kids count and martial
status along an array of income levels
}
\description{
there are m locations and n population groups, simulate the
distribution location-specific population distribution as a m by n
dataframe where the elements of the entire matrix sum jointly to 1.
}
\examples{
ls_simu_loc_demo <- ffp_evd_simu_loc_demo_main()
mt_pop_data_frac <- ls_simu_loc_demo$mt_pop_data_frac
mt_loc_pop_bernoulli <- ls_simu_loc_demo$mt_loc_pop_bernoulli
df_demo_id_desc <- ls_simu_loc_demo$df_demo_id_desc
df_location_id_desc <- ls_simu_loc_demo$df_location_id_desc
print("share of population in location and demographic cell")
print(mt_pop_data_frac)
print(paste0(
  "location- and group-specific pbernoulli",
  " from pbinom(x|n, pbernoulli_loc_group)"
))
print(mt_loc_pop_bernoulli)
print("population group id attributes")
print(df_demo_id_desc)
print("location id attributes")
print(df_location_id_desc)

}
\references{
\url{https://fanwangecon.github.io/prjenvdemo/articles/fv_rda_simu_loc_demo.html}
}
\author{
fan wang, \url{http://fanwangecon.github.io}
}
